implemented admin tool fetch user by attribute and add or remove author or editor

+++ app/routes/mediaUploadRoutes.js
@@ -109,9 +109,14 @@ module.exports = app => {
   });
 
   app.delete('/api/delete_blog_image', (req, res) => {
-    Image.findByIdAndDelete(req.query.imageId, err => {
-      if (err) {
-        console.log(err);
+    cloudinary.uploader.destroy(req.query.imagePublicId, (error, result) => {
+      // console.log(result, error);
+      if (result) {
+        Image.findByIdAndDelete(req.query.imageId, err => {
+          if (err) {
+            console.log(err);
+          }
+        });
       }
     });
   });
